src/requirements.txt
langgraph
langchain
langchain-community
langchain-databricks
databricks-langchain
pyyaml
mlflow
pandas
plotly
databricks-sdk
databricks-vectorsearch
unitycatalog-langchain
streamlit
python-dotenv
